+++ locales/cs/ui.menu.json
@@ -63,5 +63,6 @@
   "polls": "Hlasování",
   "manage": "Správa",
   "top": "Top",
-  "goals": "Cíle"
+  "goals": "Cíle",
+  "userinfo": "Uživatelské informace"
 }
