+++ locales/en/ui.menu.json
@@ -63,5 +63,6 @@
   "polls": "Polls",
   "manage": "Manage",
   "top": "Top",
-  "goals": "Goals"
+  "goals": "Goals",
+  "userinfo": "User info"
 }
